added first file README.md to test commit
This is just me testing out my first commit messages. I believe the
first line will be the subject, and these subsequent lines will be the
body of the committ message. Best practice is to keep subject less than
50 characters and each line of the body less than 72 characters. So far
I've been manually hitting enter as my text reaches a vertical line,
which seems like it is telling me to stop. I'm guessing that's why people use things like VS Code, for those little tips. This line I went way past it, and didn't go to a new line just to see what would happen.
But this is on a new line. I'm not sure how to finish this commit the
terminal says it's waiting for the editor to close the file.
I hit cmd+s to save, and I'm going to try to close now.